teams.py debugged now it is working

+++ backend/teams/team.py
@@ -14,6 +14,7 @@ import jwt
 
 # Create a team and add to the database.
 def team_create(token, team_name):
+    token = db.session.query(Token).filter_by(jwt_token=token).first()
     user = get_user_from_token(token)
     
     # Check that there are no existing teams with the specified team name.
@@ -24,6 +25,11 @@ def team_create(token, team_name):
     team = Team(name=team_name, task_master_id=user.id)
     db.session.add(team)
     db.session.commit()
+    #fill team id attribute in user object
+    user.set_team_id(team.id)
+    db.session.commit()
+    
+
     
     return {
         "team_id": team.id
@@ -31,54 +37,63 @@ def team_create(token, team_name):
 
 # Delete a team from the database.
 def team_delete(token):
+    token = db.session.query(Token).filter_by(jwt_token=token).first()
     user, team = get_user_and_team_from_token(token)
-    
+    if team is None:
+        raise AccessError('the team does not exist')
     # Only the task master of the team can delete the team.
     check_user_is_task_master(user,team)
     
     # Unassign all members from the team
-    team_members = User.query.filter_by(team_id=team.id)
+    team_members = User.query.filter_by(team_id=team.id).all()
     for team_member in team_members:
-        team_member.team_id = None
+        team_member.set_team_id(None)
+        db.session.commit()
     
     # Remove all tasks
     tasks = Task.query.filter_by(team_id=team.id)
     for task in tasks:
         db.session.query(Task).filter_by(id=task.id).delete()
+        db.session.commit()
         
     # Delete the team
-    db.session.query(Team).filter_by(id=team.id).delete()
+    team = db.session.query(Team).filter_by(id=team.id).delete()
+
     return {}
 
 # Update the task master.
 def team_update_task_master(token, new_task_master_email):
+    token = db.session.query(Token).filter_by(jwt_token=token).first()
     user, team = get_user_and_team_from_token(token)
     new_task_master = get_user_from_email(new_task_master_email)
 
     # Only the task master of the team can choose the task master.
     check_user_is_task_master(user,team)
         
-    team.task_master_id = new_task_master.id
-    
+    team.set_task_master_id(new_task_master.id)
+    db.session.commit()
+    new_task_master.set_team_id(team.id)
+    db.session.commit()
     return {
         "team_id": team.id
     }
     
 # Update the team name.
 def team_update_team_name(token, new_team_name):
+    token = db.session.query(Token).filter_by(jwt_token=token).first()
     user, team = get_user_and_team_from_token(token)
 
     # Only the task master of the team can update the team name.
     check_user_is_task_master(user,team)
-        
-    team.name = new_team_name
-    
+    team.set_name(new_team_name)
+    db.session.commit()
     return {
         "team_id": team.id
     }
     
 # Add a team member by email address to your team.
 def team_add_team_member(token, member_email_address):
+    token = db.session.query(Token).filter_by(jwt_token=token).first()
     user, team = get_user_and_team_from_token(token)
     check_user_is_task_master(user,team)
     
@@ -86,21 +101,25 @@ def team_add_team_member(token, member_email_address):
     
     # NOTE: At the moment, this function will immediately set the team member's team
     # to that of the user, regardless of whether or not the team member is already in a team.
-    team_member.team_id = user.team_id
+    team_member.set_team_id(user.team_id)
+    db.session.commit()
     return {
         "team_id": team.id
     }
     
 # Leave a team.
 def team_leave(token):
+    token = db.session.query(Token).filter_by(jwt_token=token).first()
     user = get_user_from_token(token)
-    user.team_id = None
+    user.set_team_id(None)
+    db.session.commit()
     return {
         "user_id": user.id
     }
     
 # Remove a team member from your team.
 def team_remove_member(token, member_email_address):
+    token = db.session.query(Token).filter_by(jwt_token=token).first()
     user, team = get_user_and_team_from_token(token)
     check_user_is_task_master(user,team)
 
@@ -109,7 +128,8 @@ def team_remove_member(token, member_email_address):
         raise InputError('User cannot remove themselves from a team.')
         
     team_member = get_user_from_email(member_email_address)
-    team_member.team_id = None
+    team_member.set_team_id(None)
+    db.session.commit()
     
     return {}
     
@@ -118,5 +138,7 @@ def team_remove_member(token, member_email_address):
 
 # Helper function to check if a user is a task master of a team.
 def check_user_is_task_master(user, team):
+    print(user)
+    print(team)
     if team.task_master_id != user.id:
         raise InputError('User does not have sufficient permissions.')
